refactor: 주변 매장 찾기 API에 해당하는 service, dto workspace로 이동
- 기존에는 customerService의 메서드를 worspaceService로 이동
- dto workspace 패키지로 이동
- 사용하는 메서드 workspacePort로 이동
