+++ peauty-customer-api/src/main/java/com/peauty/customer/business/customer/CustomerService.java
@@ -1,7 +1,7 @@
 package com.peauty.customer.business.customer;
 
 import com.peauty.customer.business.customer.dto.*;
-import com.peauty.customer.business.workspace.dto.GetDesignerWorkspaceResult;
+import com.peauty.customer.business.workspace.dto.GetAroundWorkspacesResult;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface CustomerService {
@@ -9,6 +9,5 @@ public interface CustomerService {
     GetCustomerProfileResult getCustomerProfile(Long customerId);
     UpdateCustomerProfileResult updateCustomerProfile(Long customerId, UpdateCustomerProfileCommand command);
     void checkCustomerNicknameDuplicated(String nickname);
-    GetAroundWorkspacesResult getAroundWorkspaces(Long customerId);
     GetDesignerBadgesForCustomerResult getDesignerBadgesByCustomer(Long designerId);
 }
