+++ peauty-customer-api/src/main/java/com/peauty/customer/business/workspace/WorkspaceService.java
@@ -1,7 +1,10 @@
 package com.peauty.customer.business.workspace;
 
+import com.peauty.customer.business.workspace.dto.GetAroundWorkspacesResult;
 import com.peauty.customer.business.workspace.dto.GetDesignerWorkspaceResult;
 
 public interface WorkspaceService {
-    GetDesignerWorkspaceResult getWorkspaceDetail(Long workspaceId);
+    GetDesignerWorkspaceResult getWorkspaceDetails(Long designerId);
+
+    GetAroundWorkspacesResult getAroundWorkspaces(Long userId);
 }
