+++ peauty-customer-api/src/main/java/com/peauty/customer/business/workspace/dto/GetAroundWorkspacesResult.java
@@ -1,4 +1,4 @@
-package com.peauty.customer.business.customer.dto;
+package com.peauty.customer.business.workspace.dto;
 
 import com.peauty.domain.customer.Customer;
